support substr and exact
